requirements.txt
requests==2.28.1
discord==2.0.0
python-dotenv==1.0.0